Change `copy(BraidingTensor)` to return a `BraidingTensor` (#2)
This changes the behaviour of copy as an instantiator for creating a TensorMap from a BraidingTensor.
The rationale is that while this does sometimes happen in Julia Base, this is always in the context of lazy wrapper types, for which it makes sense to not copy the parent and then create a new wrapper. BraidingTensor does not wrap anything, so this definition makes less sense.