[Performance] `TreeTransformer` refactor + multithreading (#251)
This PR features a change in the implementation of the treetransformers in the generic case of non-unique fusion.

The core idea is to avoid having to access data non-contiguously as much as possible, while making use of BLAS to facilitate the recoupling.
In order to achieve this, we now loop over the set of uncoupled charges and process all fusion trees that are attached to this simultaneously.
The recoupling is done through BLAS, the copying into a temporary buffer with an optimized copy and the final permutation with Strided.
Additionally, I added and centralized the implementation of multithreading for the treetransformers.

This is a result of many discussions and inspiration from @ogauthe.

Co-authored-by: Olivier Gauthe <olivier.gauthe.2011+github@polytechnique.org>